backend: split express app into shared app entrypoint
