+++ backend/src/server.ts
@@ -1,1 +1,7 @@
-import "./index";
+import app from "./app";
+
+const PORT = process.env.PORT || 5000;
+
+app.listen(PORT, () => {
+  console.log(`Server running on http://localhost:${PORT}`);
+});
